Further build optimisation:
* Exclude `internal/storage` from WASM build
* Removed 7mb+ `net/http` dependency by registering
  `jsonschema/v5/httploader` in `!wasm` file.
