+++ internal/storage/db/sqlite3/conf.go
@@ -1,6 +1,8 @@
 // Copyright 2021-2025 Zenauth Ltd.
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build !js && !wasm
+
 package sqlite3
 
 import "github.com/cerbos/cerbos/internal/storage"
